refactor(layout): move rl/ to training/rl/ (T124.4.1)
Move top-level rl/ package under training/rl/ per E124 layout cleanup
(docs/design.md). Updates import path
github.com/zerfoo/zerfoo/rl -> github.com/zerfoo/zerfoo/training/rl
in tests/parity/layer_parity_test.go.
