+++ tests/parity/layer_parity_test.go
@@ -30,11 +30,11 @@ import (
 	"github.com/zerfoo/zerfoo/layers/transformer"
 	ltranspose "github.com/zerfoo/zerfoo/layers/transpose"
 	"github.com/zerfoo/zerfoo/layers/vision"
-	"github.com/zerfoo/zerfoo/rl"
 	"github.com/zerfoo/zerfoo/synth"
 	"github.com/zerfoo/zerfoo/tabular"
 	"github.com/zerfoo/zerfoo/training/loss"
 	"github.com/zerfoo/zerfoo/training/optimizer"
+	"github.com/zerfoo/zerfoo/training/rl"
 	"github.com/zerfoo/ztensor/compute"
 	"github.com/zerfoo/ztensor/graph"
 	"github.com/zerfoo/ztensor/numeric"
